requirements.txt
# Core LLM and RAG
langchain>=0.2.0
langchain-google-genai>=1.0.0
langchain-community>=0.2.0
langchain-chroma>=0.1.0

# Vector Store
chromadb>=0.4.0

# Evaluation Framework
ragas>=0.1.0
datasets>=2.0.0

# Document Processing
beautifulsoup4>=4.12.0
requests>=2.31.0
tiktoken>=0.5.0

# Utilities
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
tqdm>=4.65.0

# For advanced RAG techniques
sentence-transformers>=2.2.0
rank-bm25>=0.2.2

